requirements.txt
paho-mqtt
bluepy
